+++ src/audiodevice.rs
@@ -505,12 +505,18 @@ pub fn new_resampler(
 
 /// Create a capture device.
 pub fn new_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
-    //let resampler = new_resampler(&conf);
+    // Use `capture_samplerate` from config if given, and resampling is enabled.
+    // Else, use `samplerate`.
     let capture_samplerate = if conf.capture_samplerate.is_some() && conf.resampler.is_some() {
         conf.capture_samplerate.unwrap()
     } else {
         conf.samplerate
     };
+    if let Some(cr) = conf.capture_samplerate {
+        if cr != conf.samplerate && conf.resampler.is_none() {
+            warn!("Resampling is disabled and capture_samplerate is different than samplerate, ignoring capture_samplerate.");
+        }
+    }
     let diff_rates = capture_samplerate != conf.samplerate;
     // Check for non-optimal resampling settings
     if !diff_rates && conf.resampler.is_some() && !conf.rate_adjust() {