+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v2.0.1
-Changes:
+Bugfixes:
 - Ignore capture_samplerate when resampling is disabled.
+- Increase Alsa device buffer sizes to avoid errors.
 
 ## v2.0.0
 New features: